The TestVORSSiteCatalog has hte logger initialized correclty now
Also now pegasus-get-sites creates path for globus location dependant on whether the auxillary jobmanager is of type fork or not.
If it is of type fork then picks up the value of GLOBUS_LOCATION variable
else it picks up the value from OSG_GRID variable i.e. GLOBUS_LOCATION is set to $OSG_GRID/globus