In order to generate different submit /dag files on basis of condor versions, we need to convert the condor version to a corresponding numeric value. Earlier there was a method CondorVersion.intValue() that was used for comparison. This was incorrect, as shown by following example 7.2.0  is 720
However
7.1.10 is 729
This has been fixed by adding a numericValue function to CondorVersion
class. This correctly computes the string value to the corresponding
long version. Appropriate error handling occurs if version is not in
the right format.

Rescue DAG Handling for recursive workflows

Vahi 17:48, 25 March 2009 (PDT)
The option of whether to do auto rescue is now triggered on basis of
the condor version . If condor version is >7.1 , then auto rescue
happens or if user explicitly set --rescue option to > 0

In PartitionAndPlan an extra debug statement was added.