Added the Regex Replica Selector that allows the user to provide the ranking order for PFN returned for various LFN.
It allows the user to specific regex expressions that can be used to
 rank various PFN's returned from the Replica Catalog  for a
 particular LFN. This replica selector selects the highest ranked  PFN
 i.e the  replica with the lowest rank value. 

The regular expressions are assigned different rank, that determine
the order in which the expressions are employed. The rank values  for
the regex can expressed in user properties using the property.

 pegasus.selector.replica.regex.rank.[value]   
  
 The value is an integer value that denotes the rank of an
 expression with a rank value of 1 being the highest rank.
 
To use this the property pegasus.selector.replica.selector must be set to value 
Regex