Added a new property pegasus.file.cleanup.scope to determine the scope of the cleanup algorithm.
By default in case of deferred planning cleanup is turned on as the
cleanup algorithm does not work across partitions.

However, in scenarios where the partitions themseleves are independant
( i.e. dont share files ), user can safely turn on cleanup. 

This can now be done by setting

pegasus.file.cleanup.scope  deferred

If the property is set to deferred, and the user wants to disable
cleanup then he can still specify --nocleanup option on command line
and that is honored.

However in case of scope fullahead,  the command line options are
ignored and always nocleanup is set