Pegasus now creates a [workflow].job.map file that links jobs in the DAG with the jobs in the DAX. The contents of the file are in netlogger format.
The [workflow] is replaced by the name of the workflow i.e. same prefix as the .dag file

In the file there are two types of events.


pegasus.job pegasus.job.map

pegasus.job - This event is for all the jobs in the DAG. The following information is associated with this event.

    * job.id the id of the job in the DAG
    * job.class an integer designating the type of the job
    * job.xform the logical transformation which the job refers to.
    * task.count the number of tasks associated with the job. This is equal to the number of pegasus.job.task events created for that job. 

pegasus.job.map - This event allows us to associate a job in the DAG with the jobs in the DAX. The following information is associated with this event.

    * task.id the id of the job in the DAG
    * task.class an integer designating the type of the job
    * task.xform the logical transformation which the job refers to. 

Also wings.jar was updated . The log messages now dont have event id or message id associated with them.