Checkin for partial fulfillment of JIRA PM-45
The DAXParser now parses dax 3.0 documents . It is able to parse the pfn's for input/output files and the executables to use.

Pegasus correctly selects the PFN's if any specified for executable or for data files. 
If an executable url does not start with /, it is taken to mean that the file needs to be staged.