For second level staging case, the worker node directories that are created in wntmp have a job specific component in the name.
The name of worker node directory created is now based on the relative
submit directory  and the jobname.

For example. if the wntmp is /tmp 
and relative submit directory is vahi/pegasus/blackdiamond/run0013

and complete jobname is vahi::preprocess:1.0

the complete worker node directory will be 
/tmp/vahi-pegasus-blackdiamond-run0013/vahi-preprocess-1.0/

This is related to JIRA feature PM-20