Kickstart - now while writing out the .in file for clustered jobs , the -w option is always added in case of clustered jobs.
NoGridStart - put in support for allowing clustered jobs if the SLS implmentation is S3. We merge the sls_* files for the clustered job into the seqexec.in file

GridStart - added a new method getWorkerNodeDirectory

AggregatedJob - added a getMethod

Related to JIRA bug improvement PM-59

This is a very specific implementation only for S3 and wont work with other SLS mechanisms.