Bug fix for PM-58
Changed the internal handling for the initialdir and remote_initialdir keys. The initialdir key is now only associated for standard universe jobs. For glidein and condor style we now associate remote_initialdir unless it is a standard universe job. 
In case of Sipht setup, however the initialdir keys are created by the Condor SLS implementation as we are using condor file transfer mechanism to get input and output data to a directory on the submit host.