Related to JIRA PM-61
Moved URL construction into the MYOSG class. It was earlier in the TestMyOSG class.  

Introduced a boolean property pegasus.catalog.site.myosg.keep.tmp.file that controls whether the temp xml file scrapped over from the MyOSG site is kept on the file system or not. By default it is deleted.

The PegasusGetSites class now fowards the vo and grid properties to the site catalog backends.