Checkin for JIRA PM-50
The entry for condor::dagman is no longer required for site local in the transformation catalog. 
Instead pegasus constructs path from the following environment variables.
CONDOR_HOME, CONDOR_LOCATION

The priority order is as follows
1) CONDOR_HOME defined in the environment
2) CONDOR_LOCATION defined in the environment
3) entry for condor::dagman for site local

There is no provision for site catalog right now, as the API in question does not have handle to the site catalog.