Related to JIRA bug PM-20 ( worker node execution ) and PM-62 ( use of SRM for data staging to worker nodes ).
Added capability to bypass first level staging if the input files in the replica catalog have a pool attribute matching the site at which a job is being run. This applies in case of worker node execution.

The cache file generated in the submit directory is the transient replica catalog. It also now has locations of where the inpute files are staged on the remote sites. Earlier it was only the files that were generated by the workflow.

The transient replica catalog handle is now stored in PegasusBag as it is required by the SLS module implementations. The Transfer and S3 SLS implementations were modified to pick up the input locations of the input data from the locations in the transient RC.