Related to JIRA bug PM-20
For worker node execution the directory on the worker node tmp was a deep directory created.  e.g if the worker node tmp is /tmp

A job maybe executed in a directory
/tmp/vahi-pegasus-blackdiamond-run0036/preprocess_ID000001 

However the above fails kickstart -W option does not create deep directories.
For the time being no deep directory is created. Now the worker node directory will be 
/tmp/vahi-pegasus-blackdiamond-run0036-preprocess_ID000001