Related to JIRA PM-45
If in a DAX 3.0 dax, only the executable is specified and the site is not specified then default entries are inserted into the transformation catalog with default architecture INTEL32::LINUX with site handle as "unknown". If url does not start with a ./ then, the type of entry is set to STATIC_BINARY

This ensures that during staging of executables these STATIC_BINARY entries are discovered in the TC and hence pegasus can successfully plan.

However this also highlights the deficiency in the format as we right now cannot specify the architecture of static binaries in the DAX. Hence, stageing of executables will work only against those sites whose sysinfo is set to INTEL32::LINUX.