Related to JIRA PM-64
If no execution sites are specified in the DAX jobs in a DAX3.0 dax, PartitionAndPlan sets the list of executions sites to the ones passed for DAX3.0 dax.

Additionally fixed, if on command line user does not specify any execution sites then all sites in site catalog are considered for planning.