Checkin for PM-51
All the dax jobs in DAX 3.0 are now rendered as SUBDAG EXTERNAL in the .dag file for the outer level workflow. The prescript for the SUBDAG jobs contain the pegasus-plan invocation that generates the sub dag's.

Additionally, changes were made to logging initialization in CPLanner. If a user specifies a file to which the logging should go using pegasus.log.* property, then a backup is created of the existing log file. This means that when a EXTERNAL SUBDAG job is retried , the pegasus-plan log files will be preserved. 

For e.g. if pegasus.log.*=/tmp/test.log  then on subsequent invocation the log will be backed to /tmp/test.log.000

The SUBDAX generator class now is called to wrap the pegasus-plan job as a SUBDAX.