Related to PM-45
Added the following attributes in the DAX 3.0 schema

User can specify the runtime for a job by specifying the runtime attribute in the job element.

User can specify the size of a file , by specifying the size attribute in the uses element. 

The DAXParser was modified to populate the runtime in the SubInfo class , and the size in the PegasusFile class. Accessor methods were added to these data classes.