Bug Fix for JIRA PM-62
In case for execution on EC2 where S3 is used, the URL to be registered in the transient RC is replaced by the S3 bucket name. The original URL contains a directory path instead of the bucket name.

This discrepancy is because the working directory on compute sites is retrived from the Site Catalog interface, that assumes always directories are created.