Related to JIRA PM-67
Britta reported that for some of the sites she tested on ,  the worker node client package was not installed on the head node. Hence the worker staging jobs failed. 

Mats  added a new OSGMM_Globus_Location_Fork attribute. 
We know use this attribute instead of OSGMM_Globus_Location to pick up the GLOBUS_LOCATION