RElated to JIRA PM-33
There was a bug whereby if PEGASUS_HOME was not set in the site catalog, the jobs were launched always via NoGridstart even though the user had worker package staging set.

This was because the GridStartFactory always loaded NoGridStart if gridstart path could not be determined from the Site Catalog. Now NoGridStart is launched by default only for mpi jobs.

Additionally, the call to handleTransferOfExecutable( SubInfo, String ) was moved to start of the function enable(SubInfo, boolean, boolean, boolean ), to ensure that a valid path to kickstart can be created before doing the sanity check.