Related to JIRA PM-68
Initial Checkin of the SeqExec Gridstart implementation for use in EC2
/ S3 environment.

S3 SLS implementation now can disable the transfer of SLS files if the
property pegasus.transfer.sls.s3.stage.sls.file  is set to false.

Added a backdoor in Kickstart implementation to set the boolean
variable that tracks whether enabling is happening during clustering
or not.