Related to JIRA PM-68
In case of clustering, the .in files for the constitutent jobs was
also created. This is now fixed in SeqExec Gridstart module.

All clustered jobs are launched in NoGridStart and then directed to
the right GridStart module.
In the case where SeqExec grid start is not used, clustering in S3
will work as implemented for JIRA PM-59