Related to jira PM-51 and JIRA PM-45
In case the dax file for the sub workflow does not exist, the label value is set to the basename prefix option passed in the dax job arguments.

An error is only thrown if both basename prefix option is not passed and the dax does not exist.

Earlier an error was always thrown if DAX did not exist.