Bug fix for JIRA PM-80
Removed the NPE that was due to the head node directory being set to
null in case of worker node execution with transfer of proxy set to
true.

Also the value of X509_USER_PROXY was not set for the compute jobs in
case of worker node execution. This is fixed.

The value was incorrectly specified for the third party transfer
jobs. That is also fixed now.