Related to JIRA PM-38
The s_t_f and w_t_f condor keys are always removed for local/scheduler universe jobs , as in certain cases the user may have these keys specified for the local site in the site catalog  to be applied for the vanilla universe jobs.

This is in case of submitting ligo workflows directly to a condor pool.