Related to JIRA PM-128
Fix for the bug.  Stageout job is only added for files for which the transfer floag is set to true. 

Additionally, this refiner does not support registration jobs as yet. 
A warning is logged if a user specifies a dax in which register flag is set true for output files.