Related to JIRA PM-131
There is now a separate Braindump Code Generator to generate a braindump file for the workflow. The Condor and Shell code generators call out to this to generate the braindump file. Braindump file is required for pegasus-run to submit the workflow , and hence for the --submit option to pegasus-plan to work in case of Shell script code generator.