Related to JIRA PM-133
Moved the internal Transformation Catalog API to 1.2

The methods that took in VDSSysInfo as a parameter now take in SysInfo
as a parameter.

Changed function names in NMI2VDSSysInfo class.

The existing TC implementation File and Database still use VDSSysInfo
internally, as that is how system information is represented in the
respective backends.