Related to JIRA PM-159
Edited comments.
Changed the order of summary : dag graph,dax graph, show-job and show-run
Changed how the error messages are displayed in execution of failure.
Displays the output of .out and .err files if -v option is set .
Displays the path to  .out and .err files and sets -keep option if -v option is not set .