JIRA PM-316
The modifyJobForFirstLevelStaging was earlier called before clustering, that could lead to stage in jobs shipping in sls files for the jobs making up a clustered job. If we ship, the SLS file via a stagein job we want only the SLS file for the merged job to be shipped.  This is now fixed.