JIRA PM-131
The generated shell script logs the debug statements to stdout instead of stderr.  Else the users maybe get confused when they use the submit option of pegasus-plan with shell code generator, as the shell script output will appear on the error log of pegasus-plan.

merging change 3375 from trunk to branch 3.0.