JIRA PM-334
CommonProperties now takes in an option for the conf properties.
PegasusProperties function signatures also changed.

Did away with the notion of passing the basename of the properties file to be picked from $PEGASUS_HOME/etc directory as we no longer support it.
The String parameter passed is now the path to the conf properties.