Fix for PM-380
Checked in the fix for the 3.0 branch.

In the 3.0 branch, the parser was parsing the various pfn's correctly, but the same TransformationCatalogEntry object was updated, as clone was not called on the instance.

Also, fixed the stack overflow issue that could occur while calling out to transformation selectors that return more than one entry.