JIRA PM-326
The argv attribute in the netlogger events is only generated if they are not null and have length > 0 to conform to stampede yang schema. 

Also the dax jobs did not have the correct dagman arguments specified in the netlogger event for the corresponding executable job. They still had what was in the DAX. This is now fixed.