JIRA PM-340
Partial commit.

We now have a numbering scheme for the submit directories for sub
workflows, with a symlink pointing to the current directory.

In case of replanning, the numbered submit directory is incremented. 

The sub workflows now refer to the properties file of the parent
workflow from the parent's submit directory. A copy is not written out
in the sub workflow submit directory when a DAX job is being planned.

Disabled the writing out of the replica store file for time being.

Right now replanning is turned on always.