JIRA PM-320 and PM-340
Changed the way how inherited-rc-files option is implemented. The
replica store is now generated in the parent workflows submit
directory i.e the directory of the workflow whose DAX contains file
locations. The sub workflow pegasus plan prescript refers to this
file.

Added a DAXReplicaStore CodeGenerator to generate the file.