JIRA PM-396
the job submit files now have two new classads
pegasus_root_wf_uuid - the uuid of root workflow 
pegasus_wf_uuid - the uuid of the workflow to which the job belongs. 

For the root workflow, root_wf_uuid and wf_uuid will be the same. 

Also pegasus-submit-dag was changed accordingly to generate these two classads for the root level workflow