JIRA PM-356
The clustered jobs after enabling had kickstart setup and cleanup env variables associated with them. That should not be the case, as we rely on the SeqExec env variables to setup and remove the worker node directories.

Also the paht to kickstart was not overriden correctly via profiles. Fixed that.