JIRA PM-328
The argument string stored in braindump file are the arguments with which the planner was invoked, not the string that is sanitized after parsing the arguments as it then contains options added internally by the planner that user did not set.

PlannerOptions has a new instance variable that stores the original arguments.