Related to JIRA PM-332
Issue: PFN file url is saved when adding an entry. However, when reading an entry the PFN protocol is trimmed out. 
Fix : Added a new property to specify whether to modify the file url or not which is set to true by default. But when used through tc-client it is set to false.