Related to JIRA PM-464
Removed the wings.jar. For 3.1 release we only have source java files
that are required for the netlogger event creation.

The EventLogMessage class now instantiates Escape to only escape
double quotes. Single quotes and \ are not escaped

For 3.2 we will move these files to the netlogger svn ( and changed
the package names appropriately) and  include that jar in Pegasus.