JIRA PM-494
Now, for executable staging in case of Condor IO also, we will refer to the
staged executable basename instead of the one corresponding to the PFN
in the Transformation Catalog. We did that earlier, since Condor IO
did not allow for renaming of file. This is not a problem any longer,
as the files are now staged via a staging site.