PM-495
The clustered jobs standard input is included in the shell
script. The version in the shell scripts is picked up from the planner
version now. also the constituent jobs have their change dir profiles
set to false, to ensure that -w or -W option does not get triggered
when being enabled by kickstart.