JIRA PM-497
Worker package staging updated for Pegasus Lite. In case of PegasusLite

- worker package is staged to the submit directory. There is a
stage_worker job that does that. However, unlike the shared filesystem
setup no untar jobs are added to the workflow 
- The worker package is then staged with job using Condor IO 
- no cleanup job is added to remove the worker package 

To enable worker package staging 
pegasus.transfer.worker.package true 

Even when worker package transfer is enabled if for a site, PEGASUS_HOME is defined then worker package is not staged from the submit directory with the job.