First integration of a policy service capability into the Pegasus transfer tool. This revision looks for the existence of an environment variable PEGASUS_POLICY_CHECKS. If set and it evaluates to "True," instead of a simple sort of the transfer list, the list is submitted to the policy service. If the environment variable is not set, processing proceeds as before.
The policy service queries require the JSON library. However, the script only attempts to load this, and the other required libraries, if the policy service is to be used. Older Python installations might not have a JSON library and therefore cannot use a policy service.

This revision tested with Python 2.6.1 as provided in a Mac OS X 10.6.8 environment.