JIRA PM-494 PM-497
In case of worker package staging and PegasusLite, the stage worker
job has null set for the staging site. This is because the stage
worker job runs on the local site, and there is no associated create
dir job for it. 
No edge is added to a create dir job for the stage worker job.