JIRA PM-538
The worker package is now staged using pegasus-transfer.

For the shared fs case, the pegasus worker executables are catalogued
into transformation catalog only after the stage worker job has been
created.
This is to ensure, that the staged path is not used for the path to
pegasus-transfer ( avoiding the chicken and egg problem).