JIRA PM-538
For transferring the worker package pegasus now uses pegasus-transfer
for all cases. 

Added a new job type STAGE_IN_WORKER_PACKAGE_JOB  that is used to
distinguish between a vanilla stagein job and stage worker job. This
is to ensure, that for any remote stage worker job scheduled,
pegasus-transfer is transferred using transfer_executable from the submit host.

This led to change in the TransferImplementation API.