Related to PM-494 , PM-499, PM-495
Making sure that the cleanup job has user executable staging set to
false. Since it inherits from the compute job, in case of executable
staging the value is set to true. Because of this, when invoked
through kickstart full path to pegasus-cleanup was not generated

this is fixed by setting executable staging attribute to false when
the cleanup job is created for PegasusLite