Adding Credential handling code.
The code checks which Credential are required by each job. It then initiates the corresponding handler and gets the location of the associated credentials.
The credential location is added to the job by setting the corresponding environment variable.

Example: A job requiring an X509 certificate will set the X509_USER_PROXY environment variable with the location of the proxy file.