Adapt the script to use a policy service if the environment is configured for it. If the environment is not configured for a policy service, then the observed behavior of the script should be the same.
To use a policy service required some changes in how the script handles its list of files to cleanup. The script still does these serially, but now the list is either sorted or passed to the policy service. On success, the policy service is notified that the files were cleaned-up.