PM-497
The execution site for the cleanup jobs is not considered while
figuring out where to deploy the worker packages, as for execution
sites a compute job will also be scheduled to be run there.

If a clean up job is scheduled to run on local site ( as in most cases
for 4.x ) , then then the pegauss installation on the submit host is
picked up.