PM-497
For Pegasus Lite when pegasus.transfer.worker.package is set to false,
we add a worker package to transfer via condor io if we can find it in
the TC. In this case, there is no stage worker job created, just that
worker package is transferred directly from the location specified in
the TC for pegasus::worker with staged set to true.

No error if we cannot find it.