Change cluster-summary line for pegasus-mpi-cluster
The new format is:

[cluster-summary stat="ok" tasks=4, succeeded=4, failed=0, extra=0, 
start="2012-05-14T15:50:12.063-07:00", duration=0.007, pid=13803, 
app="./pegasus-mpi-cluster", cores=2, utilization=0.357]

The duration field is now wall time instead of total CPU time, the cores 
field was added, and the utilization field was added. Utilization is based 
on the total number of cores, not cores-1. The cores field is meant to map 
to the multiplier_factor field in the stampede database.