Add functionality to assign a host-centric rank to each worker
Each worker is assigned a number starting at 0 relative to the host that
it is running on. The worker with the smallest global rank on each host
is given a host rank of 0, the next smallest is given host rank 1, and
so on.

This functionality will be used to implement host scripts initially.
Later we will use it to identify workers to disable for resource-based
scheduling.