PMC: Update the cluster-task record
Add the hostname for the cases where kickstart isn't used.

Add the slot # so that we can distinguish between different slots
on the same host.

Add the task name so that we can identify each task uniquely.

Also, fix a bug where, if the #@ comments are missing, the
id field came out like " id=, ". Now, if the #@ comments are missing
the id field isn't added to the cluster-task record.