PMC: Initial implementation of memory-based scheduling.
In this version we can use -m/--memory to specify the memory of a task,
and --host-memory to specify host memory, and PMC will reserve memory
for each task and ensure that hosts are not oversubscribed. The default
value for -m/--memory is 0, so if memory is not specified for one or
more tasks, then it is still possible to get OOM errors if the tasks
use too much memory.

The algorithm is greedy, meaning that it matches each ready task to the 
first slot whose host has enough memory to run it. It will keep doing
that until no matches are found (either because there are no more free
slots, there are no more ready tasks, or there are no slots capable of
running any of the ready tasks).

This version does not include support for CPU scheduling, but that will
be trivial to implement later.

It also does not include support for setrusage(), which will be 
implemented later when we figure out how to do it with mixed DAGs in
which some tasks have memory estimates and some do not.

The cluster-summary record has also been updated to include "runtime", 
which is the total cpu time consumed by all the tasks in seconds, and
"cpus", which is the total number of cpus on all hosts. In addition,
the "cores" field has been renamed to "slots" to better reflect its
real meaning.