Fix for PM-606
Added clone method for the property classes that does a deep clone.
The read.only attribute is set when connecting to replica catalog in
the planner. This ensures, that file is not written over again at end
of planning process.

In CommonProperties , setProperty function sets the property in the
internal properties object, not in System Properties.