pegasus-exitcode should look at the value of "stat" in the cluster-summary record for clustered jobs.
In the case where a PMC job terminates prematurely because it exceeded its
wall time, the value for failures may be 0, because no tasks have actually
failed yet, while the value for tasks and succeeded are both non-zero. In
this case, the only way we can tell from the cluster-summary record that
the job actually failed is to look at the value for stat, which should be
"failed" for failed jobs and "ok" for successful jobs.