PM-661
A first attempt at the  PBS code generator that generates a PBS submit
script for the workflow in the submit directory. 

The PBS script is generated whenever the PMC code generator is used.
For the time being, the nodes, ppn and walltime are hardcoded and the
user will need to change accordingly.