PM-657
Added an input directory ( --input-dir ) and an output directory option (--output-dir) for the planner. This allows the planner to read mappings for input files from an input directory . 

If , the output-dir option is set then the planner updates teh storage directory for the output site specified by the user. If none is specified , then the local site entry is updated.