Fix for montage bug that got triggered related to changes for PM-663
The bug got triggered because the map that tracks what file is deleted
by what cleanup node was no longer updated when the original stub
cleanup nodes were added.  
We had moved that update to when we created the actual clustered
cleanup graph node. 

Because of this, the algorithm ended up assigning two cleanup nodes on
the same level with the same file to be deleted. 

The fix involves checking whether a file is already deleted or not
when we create a clustered cleanup node. Also we have to ensure that
we don't create cleanup nodes that delete 0 files! We now return null
in that case and it results in the cleanup node not added.