PM-677
The planner now writes out a workflow cache file in the submit
directory. It tracks only the GET URL's on the staging sites

The planner cache used for planning purposes is never written out and
it's read only mode is set to make sure we don't write it out on close.