pegasus-kickstart: Add per-pid I/O, memory and CPU usage
These changes add one or more <proc> elements inside all of the <*job>
elements. The new <proc> elements are only available on Linux systems
with kernels >2.5.60 that support ptrace with exit events. The new
<proc> element contains information about the peak memory usage of
each child process, the start and end times of the processes, the
number of characters and bytes read and written, the utime and stime,
and the pid and parent pid. This information can be used to compute
the resource usage of a job.