kickstart: Use streams instead of buffering the xml
This is a huge commit that changes the way kickstart writes its output.
Instead of buffering the entire invocation record in memory, this change
causes kickstart to write to a standard I/O stream (FILE). Hopefully this
will alleviate some of the issues we have had with buffer sizing and quoting
of binary characters in XML.