pegasus-kickstart: Fix segfault for arg parsing
It turns out that _SC_ARG_MAX sometimes returns a very large
number on Linux (like 4 exabytes). That was causing malloc to
fail, which caused a segfault. This change sets the max
size of a single argument to 128k, which appears to be the
individual limit in Linux. If the total size of all the
arguments is over the total limit, then execve will fail, so
we don't try to catch that in the argument parser.