dag2dot: Improve handling of -f option
This fixes PM-721 by treating files and jobs as equivalent Nodes
so that transitive reduction works in the case where the DAG
contains a mix of File Nodes and Job Nodes. Non-redundant Job-Job
edges will still be rendered if the user specifies -f, but redundant
edges will be removed. If the user specifies both -f and -s, then
there will be many redundant edges in a typical workflow. Sometimes
the -f option will cause cycles in the graph (e.g. files with "inout"
linkage, or jobs with a file that is both an input and output). In
those cases the -s option must also be specified.