Mac OS X Package: Fix postinstall, python symlink
The postinstall script had a bug where it tried to install
the pegasus python library into a non-existent python version
on some systems. It was also installing it for version 2.3, which
we do not support.

Instead of using a .pth file in site-packages, which adds an entry
to the python path, it now uses a symlink, which leaves the path
unchanged.