PM-798 pegasus-analyzer by default ignores tasks that have kickstart exitcode is 0. this is for the case, where there is a clustered job in sharedfs mode that failed, and we only wwant to print tasks in the clustered job that actually failed.
So to consider a task with exitcode as 0 for printing in
pegasus-analyzer output, we add extra logic to detect the case whether
all tasks in a failed condor job succeeded. If that is the case, then
we can assume that the task with kickstart exitcode have to be
considered for printing.