Kickstart: Handle fork() edge cases in libinterpose
On a fork() we need to reinitialize libinterpose. If we fork and
exec, then we ignore the original process and just report the new
one after the exe. If we exec and then exec again, then we report
both. If we fork and never exec, then we need to report the fork.