PM-984: Propagate signals to child in Kickstart
When a user condor_rm's their job, Condor sends the job a SIGTERM. Previously
this would cause Kickstart to die. This commit changes Kickstart so that it
catches the SIGTERM and passes it on to the child instead. That way the child
dies, but not Kickstart, and Kickstart can report an invocation record for
the job to provide the user with useful debugging info. This same logic is
also applied to SIGINT and SIGQUIT.