Stop monitoring thread
When Kickstart is invoked in a way that causes a monitoring thread
to be launched, we should wait until the job is finished, and then
force the monitoring thread to write any final messages and exit.
That way if there are any messages buffered, they get sent before
Kickstart exits.