monitord: Flush sinks in main loop
This modifies monitord so that the main event loop tries to flush the database
batch before sleeping. Previously, this flushing logic could only be triggered
within the loaders, and it didn't work when there was a long period with no
events being processed. Basically, a DAGMan event needed to occur every
N seconds to trigger the N second flush in the loader. This meant that the
N second flush was not triggered in many cases, leaving a potentially large
number of events in the batch for an arbitrarily long period of time.