PM-997: Build patched pyOpenSSL on OSX
This commit restores building of pyOpenSSL on OSX. It turns out that El
Capitan does not always have OpenSSL installed as we thought. For older
versions of OSX, it should build fine with the native OpenSSL. For El Capitan
and later, it will use the homebrew OpenSSL, which will have to be installed.
pyOpenSSL has also been patched to fix a minor name conflict that prevented
it from compiling with the newer versions of OpenSSL that homebrew installs.