unpin dependencies on flask
The latest release of markupsafe has exposed a side-effect of pinning dependencies on flask, whose old versions also require an old version of jinja2. The old version of jinja2 breaks with the latest markupsafe. This means that at the moment, pegasus-wms can't be installed with pip and used as intended. The most recent version of markupsafe will be installed and thus break. 

See also https://github.com/pallets/flask/issues/4455